refactor(db): bind and read snowflakes and times directly
db::statement bound integers, text, blobs and optionals, but not the two
types every store actually has, so each store cast at every bind: about 85
static_cast<std::uint64_t> on snowflakes and a dozen
time_since_epoch().count() conversions. Each also kept its own helpers for
the optional ones: id_or_null three times, seconds_or_null twice,
snowflake_or_null, to_unix and from_unix. The casts were noise around the
SQL, and the helpers the same few lines written up to three times.

statement now binds dpp::snowflake and any std::chrono::sys_time, as whole
Unix seconds, floored as before, and get<> reads dpp::snowflake and
sys_seconds; std::optional of either works through the existing template.
The snowflake overload accepts only a snowflake, since dpp::snowflake
converts from text and would otherwise make every std::string bind
ambiguous. The nine stores drop the casts and the helpers, and
nicknames.cpp keeps to_unix only for the <t:…> timestamps it shows.

The integers written are the same, which a new test checks against the raw
column values, so existing databases read as before. util/ already includes
DPP's snowflake.h, so db/ taking it adds no new dependency.

+++ src/core/config/guild_settings.cpp
@@ -11,7 +11,7 @@
 namespace latibot::config {
 
 std::optional<std::string> guild_settings::find(dpp::snowflake guild_id, std::string_view key) const {
-    auto query = db_->prepare("SELECT value FROM guild_settings WHERE guild_id = ? AND key = ?", static_cast<std::uint64_t>(guild_id), key);
+    auto query = db_->prepare("SELECT value FROM guild_settings WHERE guild_id = ? AND key = ?", guild_id, key);
     if (!query.step()) {
         return std::nullopt;
     }
@@ -78,7 +78,7 @@ void guild_settings::set(dpp::snowflake guild_id, std::string_view key, std::str
     db_->prepare(
            "INSERT INTO guild_settings (guild_id, key, value) VALUES (?, ?, ?) "
            "ON CONFLICT(guild_id, key) DO UPDATE SET value = excluded.value",
-           static_cast<std::uint64_t>(guild_id), key, value)
+           guild_id, key, value)
         .run();
 }
 
@@ -97,14 +97,14 @@ void guild_settings::set_real(dpp::snowflake guild_id, std::string_view key, dou
 bool guild_settings::erase(dpp::snowflake guild_id, std::string_view key) {
     const auto guard = db_->lock();
 
-    db_->prepare("DELETE FROM guild_settings WHERE guild_id = ? AND key = ?", static_cast<std::uint64_t>(guild_id), key).run();
+    db_->prepare("DELETE FROM guild_settings WHERE guild_id = ? AND key = ?", guild_id, key).run();
     return db_->changes() > 0;
 }
 
 std::map<std::string, std::string, std::less<>> guild_settings::all(dpp::snowflake guild_id) const {
     std::map<std::string, std::string, std::less<>> settings;
 
-    auto query = db_->prepare("SELECT key, value FROM guild_settings WHERE guild_id = ?", static_cast<std::uint64_t>(guild_id));
+    auto query = db_->prepare("SELECT key, value FROM guild_settings WHERE guild_id = ?", guild_id);
     while (query.step()) {
         settings.emplace(query.get<std::string>(0), query.get<std::string>(1));
     }
